单机版本完结

+++ admin-core/src/main/java/com/roy/sqwaimai/bean/constant/Const.java
@@ -23,7 +23,7 @@ public interface Const {
     /**
      * 超级管理员角色id
      */
-    Integer ADMIN_ROLE_ID = 1;
+    Long ADMIN_ROLE_ID = 1L;
 
     /**
      * 接口文档的菜单名
